Issue #15 : Log room system
Add a log file, when a user connects to the room he receives all the
conversation of the room into his chat.

Add the new entry at the end of the file, that was more difficult to add
a new entry at the beginning of it.

You may need a folder data under "current version"

TODO rename the file with the room name